Tách checkout thành component con
