+++ src/app/checkout/page.tsx
@@ -1,14 +1,8 @@
 import React, { useState, useEffect, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { ShippingForm } from '@/components/checkout/ShippingForm';
-import { PaymentSelection } from '@/components/checkout/PaymentSelection';
-import { OrderConfirmation } from '@/components/checkout/OrderConfirmation';
-import CheckoutSummary from '@/components/checkout/CheckoutSummary';
 import { ShippingInfo, PaymentMethod, CheckoutState, OrderSummary, DiscountCode } from '@/types/checkout';
 import { CartItem } from '@/types/cart';
-import { StepIndicator, CheckoutStep } from '@/components/checkout/StepIndicator';
-import { OrderProcessingSkeleton } from '@/components/ui/skeleton';
-import { motion, AnimatePresence } from 'framer-motion';
+import { CheckoutComponent } from '@/components/checkout/checkout';
 
 // Dữ liệu giả lập giỏ hàng (trong thực tế, bạn sẽ lấy từ Redux store hoặc context)
 const mockCartItems: CartItem[] = [
@@ -190,113 +184,20 @@ export default function CheckoutPage() {
     }, 2000); // Giả lập 2 giây xử lý
   };
 
-  // Component hiển thị theo bước hiện tại
-  const renderCurrentStep = () => {
-    switch (checkoutState.currentStep) {
-      case 'shipping':
-        return (
-          <ShippingForm
-            shippingInfo={checkoutState.shippingInfo}
-            setShippingInfo={handleUpdateShippingInfo}
-            onNext={() => handleNextStep('payment')}
-          />
-        );
-      case 'payment':
-        return (
-          <PaymentSelection
-            selectedMethod={checkoutState.paymentMethod}
-            setPaymentMethod={handleUpdatePaymentMethod}
-            onNext={() => handleNextStep('confirmation')}
-            onBack={() => handlePreviousStep('shipping')}
-          />
-        );
-      case 'confirmation':
-        return (
-          <OrderConfirmation
-            shippingInfo={checkoutState.shippingInfo}
-            paymentMethod={checkoutState.paymentMethod!}
-            orderSummary={orderSummary}
-            onBack={() => handlePreviousStep('payment')}
-            onPlaceOrder={handlePlaceOrder}
-          />
-        );
-      default:
-        return null;
-    }
-  };
-
-  // Component hiển thị tiến trình thanh toán
-  const renderCheckoutProgress = () => {
-    const steps = [
-      { key: 'shipping' as CheckoutStep, label: 'Thông tin giao hàng' },
-      { key: 'payment' as CheckoutStep, label: 'Phương thức thanh toán' },
-      { key: 'confirmation' as CheckoutStep, label: 'Xác nhận đơn hàng' },
-    ];
-
-    // Cho phép người dùng quay lại các bước đã hoàn thành
-    const handleStepClick = (step: CheckoutStep) => {
-      // Chỉ cho phép quay lại các bước đã hoàn thành, không cho nhảy cóc
-      const currentStepIndex = steps.findIndex((s) => s.key === checkoutState.currentStep);
-      const clickedStepIndex = steps.findIndex((s) => s.key === step);
-      
-      if (clickedStepIndex <= currentStepIndex) {
-        setCheckoutState((prev) => ({
-          ...prev,
-          currentStep: step
-        }));
-      }
-    };
-
-    return (
-      <StepIndicator 
-        steps={steps} 
-        currentStep={checkoutState.currentStep as CheckoutStep}
-        onStepClick={handleStepClick}
-        allowNavigation={true}
-      />
-    );
-  };
-
+  // Sử dụng component checkout thay vì render trực tiếp UI
   return (
-    <div className="container mx-auto py-8 px-4">
-      <h1 className="text-2xl md:text-3xl font-bold text-center mb-6">Thanh toán</h1>
-      
-      {renderCheckoutProgress()}
-      
-      <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
-        <div className="lg:col-span-2">
-          {isProcessing ? (
-            <motion.div
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              exit={{ opacity: 0 }}
-              transition={{ duration: 0.3 }}
-            >
-              <OrderProcessingSkeleton />
-            </motion.div>
-          ) : (
-            <AnimatePresence mode="wait">
-              <motion.div
-                key={checkoutState.currentStep}
-                initial={{ opacity: 0, x: 10 }}
-                animate={{ opacity: 1, x: 0 }}
-                exit={{ opacity: 0, x: -10 }}
-                transition={{ duration: 0.2 }}
-              >
-                {renderCurrentStep()}
-              </motion.div>
-            </AnimatePresence>
-          )}
-        </div>
-        <div className="lg:col-span-1">
-          <CheckoutSummary 
-            items={cartItems} 
-            subtotal={subtotal} 
-            discountCode={checkoutState.discountCode} 
-            onApplyDiscount={handleApplyDiscount} 
-          />
-        </div>
-      </div>
-    </div>
+    <CheckoutComponent 
+      cartItems={cartItems}
+      subtotal={subtotal}
+      checkoutState={checkoutState}
+      orderSummary={orderSummary}
+      isProcessing={isProcessing}
+      onUpdateShippingInfo={handleUpdateShippingInfo}
+      onUpdatePaymentMethod={handleUpdatePaymentMethod}
+      onApplyDiscount={handleApplyDiscount}
+      onNextStep={handleNextStep}
+      onPreviousStep={handlePreviousStep}
+      onPlaceOrder={handlePlaceOrder}
+    />
   );
 }
